+++ projects/js-packages/number-formatters/.gitattributes
@@ -9,4 +9,4 @@ node_modules      export-ignore
 # Files to exclude from the mirror repo
 /changelog/**     production-exclude
 /jest.config.cjs  production-exclude
-/tests/**         production-exclude
+**/test/**         production-exclude
