+++ projects/js-packages/number-formatters/src/number-format-currency/index.ts
@@ -2,11 +2,7 @@ import debugFactory from 'debug';
 import { FALLBACK_CURRENCY } from '../constants.js';
 import { getCachedFormatter } from '../get-cached-formatter.js';
 import { defaultCurrencyOverrides } from './currencies.js';
-import type {
-	CurrencyOverride,
-	CurrencyObject,
-	NumberFormatCurrencyParams,
-} from '../../types/index.js';
+import type { CurrencyOverride, CurrencyObject, NumberFormatCurrencyParams } from '../types.js';
 
 const debug = debugFactory( 'number-formatters:number-format-currency' );
