+++ projects/js-packages/number-formatters/src/test/create-number-formatters.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it, jest } from '@jest/globals';
-import createNumberFormatters from '../src/create-number-formatters';
+import createNumberFormatters from '../create-number-formatters';
 
 type IntlType = typeof Intl & {
 	NumberFormat: typeof Intl.NumberFormat;
