+++ projects/js-packages/number-formatters/src/test/get-cached-formatter.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from '@jest/globals';
-import { getCachedFormatter } from '../src/get-cached-formatter.js';
+import { getCachedFormatter } from '../get-cached-formatter.js';
 
 describe( 'getCachedFormatter()', () => {
 	it( 'should return a cached formatter when one exists', () => {
