+++ projects/js-packages/number-formatters/src/test/number-format.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from '@jest/globals';
-import { numberFormat, numberFormatCompact } from '../src/number-format.js';
+import { numberFormat, numberFormatCompact } from '../number-format.js';
 
 describe( 'numberFormat()', () => {
 	it( 'should format number with default options', () => {
