+++ projects/js-packages/number-formatters/tsconfig.json
@@ -1,6 +1,6 @@
 {
 	"extends": "jetpack-js-tools/tsconfig.tsc.json",
-	"include": [ "./index.ts" ],
+	"include": [ "src/index.ts" ],
 	"compilerOptions": {
 		"typeRoots": [ "./node_modules/@types/" ],
 		"sourceMap": false,
